Detect_missing_libraries on build
The purpouse of this PR is to remove the possiblitityto use the detect_missing_libraries from the config toml file, because it does not makes sense to havie it there.
The way this flags works, is to show a list of libraries that are missing, and it does not output any bytecode, but there some other scenarios where this could happen, so the best way to fix it is to remove this from all the commands (script, test, create) which is the same as disable it by default in all places and enabling it only for the build.
finally, if there is someone still using this flag in the config, we show a warning mentioning that this is not the way to go.

fix